add mappers validation swagger logging

+++ src/main/java/me/ssagan/springdatajpa/service/AuthorServiceImpl.java
@@ -5,47 +5,84 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import lombok.RequiredArgsConstructor;
+import me.ssagan.springdatajpa.SpringDataJpaApplication;
 import me.ssagan.springdatajpa.dto.AuthorCreateDto;
 import me.ssagan.springdatajpa.dto.AuthorDto;
 import me.ssagan.springdatajpa.dto.AuthorWithBooksDto;
-import me.ssagan.springdatajpa.dto.BookWithGenreDto;
+import me.ssagan.springdatajpa.dto.mapper.AuthorCreateDtoMapper;
+import me.ssagan.springdatajpa.dto.mapper.AuthorDtoMapper;
+import me.ssagan.springdatajpa.dto.mapper.AuthorWithBooksDtoMapper;
 import me.ssagan.springdatajpa.entity.Author;
 import me.ssagan.springdatajpa.repository.AuthorRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class AuthorServiceImpl implements AuthorService {
     final AuthorRepository repository;
+    final AuthorCreateDtoMapper authorCreateDtoMapper;
+    final AuthorWithBooksDtoMapper authorWithBooksDtoMapper;
+    final AuthorDtoMapper authorDtoMapper;
+    static final Logger log = LoggerFactory.getLogger(SpringDataJpaApplication.class);
 
     @Override
     public List<AuthorDto> getAllAuthors() {
-        return repository.findAll().stream().map(this::convertToAuthorDto).toList();
+        log.info("Getting all authors");
+        return repository.findAll().stream().map(authorDtoMapper::toDto).toList();
     }
 
     @Override
     public AuthorWithBooksDto getAuthorById(Long id) {
-        Author author = repository.findById(id).orElseThrow();
-        return convertToDto(author);
+        log.info("Try to find author by id {}", id);
+        Optional<Author> authorOptional = repository.findById(id);
+        if (authorOptional.isPresent()) {
+            AuthorWithBooksDto dto = authorWithBooksDtoMapper.toDto(authorOptional.get());
+            log.info("Author: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Author with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorWithBooksDto getAuthorByName(String name) {
-        Author author = repository.findAuthorByName(name).orElseThrow();
-        return convertToDto(author);
+        log.info("Try to find author by name {}", name);
+        Optional<Author> authorOptional = repository.findAuthorByName(name);
+        if(authorOptional.isPresent()){
+            AuthorWithBooksDto dto = authorWithBooksDtoMapper.toDto(authorOptional.get());
+            log.info("Author: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorWithBooksDto getAuthorByNameBySql(String name) {
-        Author author = repository.findAuthorByNameBySql(name).orElseThrow();
-        return convertToDto(author);
+        log.info("Try to find author by name {}", name);
+        Optional<Author> authorOptional = repository.findAuthorByNameBySql(name);
+        if(authorOptional.isPresent()){
+            AuthorWithBooksDto dto = authorWithBooksDtoMapper.toDto(authorOptional.get());
+            log.info("Author: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorWithBooksDto getAuthorByNameByCriteria(String name) {
+        log.info("Try to find author by name {}", name);
         Specification<Author> specification = Specification.where(new Specification<Author>() {
             @Override
             public Predicate toPredicate(Root<Author> root,
@@ -55,39 +92,81 @@ public class AuthorServiceImpl implements AuthorService {
             }
         });
 
-        Author author = repository.findOne(specification).orElseThrow();
-        return convertToDto(author);
+        Optional<Author> authorOptional = repository.findOne(specification);
+        if(authorOptional.isPresent()){
+            AuthorWithBooksDto dto = authorWithBooksDtoMapper.toDto(authorOptional.get());
+            log.info("Author: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorDto createAuthor(AuthorCreateDto dto) {
-        Author author = convertToEntity(dto);
-        Author savedAuthor = repository.save(author);
-        return convertToAuthorDto(savedAuthor);
+        log.info("Try to create author: {}", dto.toString());
+        Author author = authorCreateDtoMapper.toEntity(dto);
+        try {
+            Author savedAuthor = repository.save(author);
+            AuthorDto saved_dto = authorDtoMapper.toDto(savedAuthor);
+            log.info("Author saved: {}", saved_dto.toString());
+            return saved_dto;
+        } catch (RuntimeException e) {
+            log.error("Author not saved: " + e.toString());
+            throw new RuntimeException(e.toString());
+        }
     }
 
     @Override
     public AuthorDto updateAuthor(Long id, AuthorCreateDto dto) {
-        Author author = repository.findById(id).orElseThrow();
-        author.setName(dto.getName());
-        author.setSurname(dto.getSurname());
-        Author savedAuthor = repository.save(author);
-        return convertToAuthorDto(savedAuthor);
+        log.info("Try to update author id: {}, values: {}", id, dto.toString());
+        Optional<Author> authorOptional = repository.findById(id);
+        if (authorOptional.isPresent()) {
+            Author author = authorOptional.get();
+            author.setName(dto.getName());
+            author.setSurname(dto.getSurname());
+            try {
+                Author savedAuthor = repository.save(author);
+                AuthorDto saved_dto = authorDtoMapper.toDto(savedAuthor);
+                log.info("Author updated: {}", saved_dto.toString());
+                return authorDtoMapper.toDto(savedAuthor);
+            } catch (RuntimeException e) {
+                log.error("Author not saved: " + e.toString());
+                throw new RuntimeException(e.toString());
+            }
+        } else {
+            log.error("Author with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public void deleteAuthor(Long id) {
-        repository.deleteById(id);
+        log.info("Try to delete author id: {}", id);
+        Optional<Author> authorOptional = repository.findById(id);
+        if (authorOptional.isPresent()) {
+            try {
+                repository.deleteById(id);
+                log.info("Author deleted id: {}", id);
+            } catch (RuntimeException e) {
+                log.error("Author not deleted: " + e.toString());
+                throw new RuntimeException(e.toString());
+            }
+        } else {
+            log.error("Author with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
-    private Author convertToEntity(AuthorCreateDto dto) {
+    /*private Author convertToEntity(AuthorCreateDto dto) {
         return Author.builder()
                 .name(dto.getName())
                 .surname(dto.getSurname())
                 .build();
-    }
+    }*/
 
-    private AuthorWithBooksDto convertToDto(Author author) {
+    /*private AuthorWithBooksDto convertToDto(Author author) {
         List<BookWithGenreDto> bookDtoList = author.getBooks()
                 .stream()
                 .map(book -> BookWithGenreDto
@@ -104,14 +183,14 @@ public class AuthorServiceImpl implements AuthorService {
                 .surname(author.getSurname())
                 .books(bookDtoList)
                 .build();
-    }
+    }*/
 
-    private AuthorDto convertToAuthorDto(Author author) {
+    /*private AuthorDto convertToAuthorDto(Author author) {
         return AuthorDto
                 .builder()
                 .id(author.getId())
                 .name(author.getName())
                 .surname(author.getSurname())
                 .build();
-    }
+    }*/
 }
