+++ src/main/java/me/ssagan/springdatajpa/service/BookServiceImpl.java
@@ -5,17 +5,22 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import lombok.RequiredArgsConstructor;
+import me.ssagan.springdatajpa.SpringDataJpaApplication;
 import me.ssagan.springdatajpa.dto.*;
+import me.ssagan.springdatajpa.dto.mapper.*;
 import me.ssagan.springdatajpa.entity.Author;
 import me.ssagan.springdatajpa.entity.Book;
 import me.ssagan.springdatajpa.entity.Genre;
 import me.ssagan.springdatajpa.repository.AuthorRepository;
 import me.ssagan.springdatajpa.repository.BookRepository;
 import me.ssagan.springdatajpa.repository.GenreRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -23,37 +28,68 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class BookServiceImpl implements BookService {
-
     final BookRepository repository;
     final GenreRepository genreRepository;
     final AuthorRepository authorRepository;
+    final BookDtoMapper bookDtoMapper;
+    final BookWithGenreDtoMapper bookWithGenreDtoMapper;
+    final BookWithGenreAndAuthorsDtoMapper bookWithGenreAndAuthorsDtoMapper;
+    final BookCreateDtoMapper bookCreateDtoMapper;
+    final BookCreateByStringsDtoMapper bookCreateByStringsDtoMapper;
+    static final Logger log = LoggerFactory.getLogger(SpringDataJpaApplication.class);
 
     @Override
     public List<BookWithGenreAndAuthorsDto> gelAllBooks() {
-        List<Book> bookList= repository.findAll();
-        return bookList.stream().map(this::convertToBookWithGenreAndAuthorsDtoDto).toList();
+        log.info("Getting all books");
+        List<Book> bookList = repository.findAll();
+        return bookList.stream().map(bookWithGenreAndAuthorsDtoMapper::toDto).toList();
     }
 
     @Override
     public BookWithGenreAndAuthorsDto getBookById(Long id) {
-        Book book = repository.findById(id).orElseThrow();
-        return convertToBookWithGenreAndAuthorsDtoDto(book);
+        log.info("Try to find book by id {}", id);
+        Optional<Book> bookOptional = repository.findById(id);
+        if (bookOptional.isPresent()) {
+            BookWithGenreAndAuthorsDto dto = bookWithGenreAndAuthorsDtoMapper.toDto(bookOptional.get());
+            log.info("Book: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Book with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public BookWithGenreDto getBookByName(String name) {
-        Book book = repository.findBookByName(name).orElseThrow();
-        return convertToDto(book);
+        log.info("Try to find book by name {}", name);
+        Optional<Book> bookOptional = repository.findBookByName(name);
+        if (bookOptional.isPresent()) {
+            BookWithGenreDto dto = bookWithGenreDtoMapper.toDto(bookOptional.get());
+            log.info("Book: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Book with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public BookWithGenreDto getBookByNameBySql(String name) {
-        Book book = repository.findBookByNameBySql(name).orElseThrow();
-        return convertToDto(book);
+        log.info("Try to find book by name {}", name);
+        Optional<Book> bookOptional = repository.findBookByNameBySql(name);
+        if (bookOptional.isPresent()) {
+            BookWithGenreDto dto = bookWithGenreDtoMapper.toDto(bookOptional.get());
+            log.info("Book: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Book with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public BookWithGenreDto getByNameByCriteria(String name) {
+        log.info("Try to find book by name {}", name);
         Specification<Book> specification = Specification.where(new Specification<Book>() {
             @Override
             public Predicate toPredicate(Root<Book> root,
@@ -62,44 +98,89 @@ public class BookServiceImpl implements BookService {
                 return cb.equal(root.get("name"), name);
             }
         });
-        Book book = repository.findOne(specification).orElseThrow();
-        return convertToDto(book);
+        Optional<Book> bookOptional = repository.findOne(specification);
+        if (bookOptional.isPresent()) {
+            BookWithGenreDto dto = bookWithGenreDtoMapper.toDto(bookOptional.get());
+            log.info("Book: {}", dto.toString());
+            return dto;
+        } else {
+            log.error("Book with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public BookDto createBook(BookCreateDto dto) {
-        Book book = convertToEntity(dto);
-        Book savedBook = repository.save(book);
-        return convertToBookDto(savedBook);
+        log.info("Try to create book: {}", dto.toString());
+        Book book = bookCreateDtoMapper.toEntity(dto);
+        try {
+            Book savedBook = repository.save(book);
+            BookDto saved_dto = bookDtoMapper.toDto(savedBook);
+            log.info("Book saved: {}", saved_dto.toString());
+            return saved_dto;
+        } catch (RuntimeException e) {
+            log.error("Book not saved: " + e.toString());
+            throw new RuntimeException(e.toString());
+        }
     }
 
     @Override
     public BookDto updateBook(Long id, BookCreateDto dto) {
-        Book book = repository.findById(id).orElseThrow();
-        book.setName(dto.getName());
-        book.setGenre(Genre.builder().id(dto.getGenre_id()).build());
-        book.setAuthors(dto.getAuthors().stream().map(author -> Author
-                .builder()
-                .id(author.getId())
-                .build()).collect(Collectors.toSet()));
-        Book savedBook = repository.save(book);
-        return convertToBookDto(savedBook);
+        log.info("Try to update book id: {}, values: {}", id, dto.toString());
+        Optional<Book> bookOptional = repository.findById(id);
+        if (bookOptional.isPresent()) {
+            Book book = bookOptional.get();
+            book.setName(dto.getName());
+            book.setGenre(Genre.builder().id(dto.getGenre_id()).build());
+            book.setAuthors(dto.getAuthors().stream().map(author -> Author
+                    .builder()
+                    .id(author.getId())
+                    .build()).collect(Collectors.toSet()));
+            try {
+                Book savedBook = repository.save(book);
+                BookDto saved_dto = bookDtoMapper.toDto(savedBook);
+                log.info("Book updated: {}", saved_dto.toString());
+                return saved_dto;
+            } catch (RuntimeException e) {
+                log.error("Author not saved: " + e.toString());
+                throw new RuntimeException(e.toString());
+            }
+        } else {
+            log.error("Book with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public void deleteBook(Long id) {
-        repository.deleteById(id);
+        log.info("Try to delete book id: {}", id);
+        Optional<Book> bookOptional = repository.findById(id);
+        if (bookOptional.isPresent()) {
+            try {
+                repository.deleteById(id);
+                log.info("Book deleted id: {}", id);
+            } catch (RuntimeException e) {
+                log.error("Book not deleted: " + e.toString());
+                throw new RuntimeException(e.toString());
+            }
+        } else {
+            log.error("Book with id {} not found", id);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public BookDto createBookByStrings(BookCreateByStringsDto dto) {
-        Book book = convertToEntity(dto);
+        log.info("Try to create book: {}", dto.toString());
+
+        Book book = bookCreateByStringsDtoMapper.toEntity(dto);
 
         Genre genre;
         Optional<Genre> genreOptional = genreRepository.findGenreByName(book.getGenre().getName());
         if (genreOptional.isPresent()) {
             genre = genreOptional.get();
         } else {
+            log.info("Try to create genre: {}", dto.getGenre().toString());
             genre = genreRepository.save(Genre
                     .builder()
                     .name(book.getGenre().getName())
@@ -117,6 +198,7 @@ public class BookServiceImpl implements BookService {
                     if (authorOptional.isPresent()) {
                         author = authorOptional.get();
                     } else {
+                        log.info("Try to create author: {}", authorCreateDto.toString());
                         author = authorRepository.save(Author
                                 .builder()
                                 .name(authorCreateDto.getName())
@@ -127,11 +209,17 @@ public class BookServiceImpl implements BookService {
                 })
                 .collect(Collectors.toSet());
         book.setAuthors(authorSet);
-        Book savedBook = repository.save(book);
-        return convertToBookDto(savedBook);
+        try {
+            Book savedBook = repository.save(book);
+            BookDto saved_dto = bookDtoMapper.toDto(savedBook);
+            return saved_dto;
+        } catch (RuntimeException e) {
+            log.error("Book not saved: " + e.toString());
+            throw new RuntimeException(e.toString());
+        }
     }
 
-    private Book convertToEntity(BookCreateDto dto) {
+    /*private Book convertToEntity(BookCreateDto dto) {
         Set<Author> authorSet = dto.getAuthors()
                 .stream()
                 .map(author -> Author
@@ -149,9 +237,9 @@ public class BookServiceImpl implements BookService {
                 .genre(genre)
                 .authors(authorSet)
                 .build();
-    }
+    }*/
 
-    private Book convertToEntity(BookCreateByStringsDto dto) {
+    /*private Book convertToEntity(BookCreateByStringsDto dto) {
         Set<Author> authorSet = dto.getAuthors()
                 .stream()
                 .map(author -> Author
@@ -170,26 +258,26 @@ public class BookServiceImpl implements BookService {
                 .genre(genre)
                 .authors(authorSet)
                 .build();
-    }
+    }*/
 
-    BookWithGenreDto convertToDto(Book book) {
+    /*BookWithGenreDto convertToDto(Book book) {
         return BookWithGenreDto
                 .builder()
                 .id(book.getId())
                 .name(book.getName())
                 .genre(book.getGenre().getName())
                 .build();
-    }
+    }*/
 
-    BookDto convertToBookDto(Book book) {
+    /*BookDto convertToBookDto(Book book) {
         return BookDto
                 .builder()
                 .id(book.getId())
                 .name(book.getName())
                 .build();
-    }
+    }*/
 
-    BookWithGenreAndAuthorsDto convertToBookWithGenreAndAuthorsDtoDto(Book book) {
+    /*BookWithGenreAndAuthorsDto convertToBookWithGenreAndAuthorsDtoDto(Book book) {
         List<AuthorDto> authorDtoList = book
                 .getAuthors()
                 .stream()
@@ -207,5 +295,5 @@ public class BookServiceImpl implements BookService {
                 .genre(book.getGenre().getName())
                 .authors(authorDtoList)
                 .build();
-    }
+    }*/
 }
